refactor(pyrovelocity): move logging module to root Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
