+++ src/pyrovelocity/flytezen/workflows/lrwine.py
@@ -1,13 +1,20 @@
-from dataclasses import asdict, make_dataclass
+from dataclasses import asdict
+from dataclasses import make_dataclass
 from datetime import timedelta
 from pprint import pformat
-from typing import Any, Dict, Optional, Tuple, Type
+from typing import Any
+from typing import Dict
+from typing import Optional
+from typing import Tuple
+from typing import Type
 
 import joblib
 import pandas as pd
 
 # from dataclasses_json import DataClassJsonMixin as DataClassJSONMixin
-from flytekit import Resources, task, workflow
+from flytekit import Resources
+from flytekit import task
+from flytekit import workflow
 from flytekit.extras.accelerators import T4
 from flytekit.types.file import JoblibSerializedFile
 from mashumaro.mixins.json import DataClassJSONMixin
@@ -15,7 +22,8 @@ from sklearn.datasets import load_wine
 from sklearn.linear_model import LogisticRegression
 
 from pyrovelocity.flytezen.configuration import create_dataclass_from_callable
-from pyrovelocity.flytezen.logging import configure_logging
+from pyrovelocity.logging import configure_logging
+
 
 logger = configure_logging("pyrovelocity.flytezen.workflows.lrwine")
